+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ unreleased
 * fix: Provide content links in the Content Expiry changelist and csv file export
 * fix: Content Expiry changelist filter dates should be for the future not the past
 * fix: CSV export Dates are not formatting correctly in Excel
+* fix: CSV export external urls used for for the url column
 
 0.0.1 (2021-10-04)
 ==================